fraud-detection-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
torch==2.3.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
python-dotenv==1.0.1